Change plugins directory to same level as project. Add basic Readme.md
Previously the SDRunoPlugin_Fran directory was under the plugins directory with the other example plugins. Since the plugins directory is also a git repository, the project files were changed to move the SDRunoPlugin_Fran directory to the same level as the plugins directory. This allows updating both more cleanly.

Added a basic Readme.md file with basic information and some compile notes.