Use new SP1 Spectrum power limits in annotation display
The new Plugin API Version 2 provides the frequency and power limiits of the SP1 Spectrum window. These changes use the new API to set the top and bottom limits of the annotations.

Update code so Plugin API Version 2 is supported. Compile-time error and warning messages were added.
Plugin API Version 1 is supported with reduced functionality. The top and bottom limits of the annotations are fixed in Plugin API Version 1

Update Version to 0.5.  This is the first release provided to SDRplay for the community plugin installer.